+++ src/apn/commons/_karatsuba_mul.c
@@ -1,5 +1,5 @@
 #include "_hidden_mul.h"
-#include "../apn_thresholds.h"
+#include "apn_thresholds.h"
 
 /*
 *	TO DO:
