+++ src/apn/commons/apn_cpy.c
@@ -1,4 +1,4 @@
-#include "../../../../include/apac.h"
+#include "../../../include/apac.h"
 
 extern __apac_cpu_params curr_cpu;
