+++ src/apn/x64/_sub_n_one.asm
@@ -30,25 +30,27 @@ _sub_n_one PROC FRAME
     ; pushing rbp sets it to a 16-byte boundary
     ; as per x64 calling convention of Microsoft ABI
 
-    xor     rax, rax    ; accumulator
     xor     r10, r10    ; counter/indexer
     mov     r11, r8     ; temp_size
 
-    mov     rax, QWORD PTR [rdx + r10*8]
+    mov     rax, QWORD PTR [rdx]
     sub     rax, r9                         ; sub val
-    mov     QWORD PTR [rcx + r10*8], rax
-    inc     r10
+    mov     QWORD PTR [rcx], rax
+    
+    lea     rdx, [rdx + 8]
+    lea     rcx, [rcx + 8]
     dec     r11
     jz      end_of_func
 
 propagate_carry:
 
     jnc     copy_remaining
-    mov     rax, QWORD PTR [rdx + r10*8]
+    mov     rax, QWORD PTR [rdx]
     sbb     rax, 0
-    mov     QWORD PTR [rcx + r10*8], rax
+    mov     QWORD PTR [rcx], rax
 
-    inc     r10
+    lea     rdx, [rdx + 8]
+    lea     rcx, [rcx + 8]
     dec     r11
     jnz     propagate_carry
     jmp     end_of_func
@@ -60,15 +62,11 @@ copy_remaining:
     ; allocate shadow space
 
     sub     rsp, 32
-    
-    lea     rcx, [rcx + r10*8]
-    lea     rdx, [rdx + r10*8]
     mov     r8, r11
 
     call    apn_cpy
     
-    add     rsp, 32
-    clc                 
+    add     rsp, 32                
 
 end_of_func:
